Two endpoints are implemented, one that returns all available movies and the other that returns detailed information about a given movie.
Implementation of unit tests for the getMovie and allMovies endpoints.

The mapping class is changed to use Mapstruct.

The mapping class is changed to use Mapstruct.

MovieService unit tests are implemented.

MovieService unit tests are implemented.

The mapper is configured by annotating the @Mapper annotation with (componentModel = "spring") and the AppConfig configuration class is removed.

The mapper is configured by annotating the @Mapper annotation with (componentModel = "spring") and the AppConfig configuration class is removed.

Implementation of unit tests of the controller using MockMvc.

Configuration intellij_codestyle.

The names of the service and its respective test are converted to plural.

The names of the controller and its respective test are converted to plural.

Mapper observations were corrected.

The formatting utility file is deleted.

MovieMapper is formatted.

Test getMovieNotFound() was fixed and now runs correctly.